requirements.txt
pypdf2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.12.0
streamlit==1.28.0
numpy==1.24.3